Clean up wasi-http interface
