+++ examples/wasi-http-rust-async/src/lib.rs
@@ -1,35 +1,32 @@
-use {
-    anyhow::{bail, Result},
-    futures::{stream, FutureExt, SinkExt, StreamExt, TryStreamExt},
-    sha2::{Digest, Sha256},
-    spin_sdk::wasi_http::executor,
-    spin_sdk::wasi_http::types::{
-        Fields, IncomingRequest, Method, OutgoingBody, OutgoingRequest, OutgoingResponse,
-        ResponseOutparam, Scheme,
-    },
-    spin_sdk::wasi_http_component,
-    std::str,
-    url::Url,
+use anyhow::{bail, Result};
+use futures::{stream, SinkExt, StreamExt, TryStreamExt};
+use spin_sdk::wasi_http::send;
+use spin_sdk::wasi_http::{
+    Fields, IncomingRequest, Method, OutgoingBody, OutgoingRequest, OutgoingResponse,
+    ResponseOutparam, Scheme,
 };
+use spin_sdk::wasi_http_component;
+use url::Url;
 
 const MAX_CONCURRENCY: usize = 16;
 
 #[wasi_http_component]
 async fn handle_request(request: IncomingRequest, response_out: ResponseOutparam) -> Result<()> {
-    let method = request.method();
-    let path = request.path_with_query();
     let headers = request.headers().entries();
 
-    match (method, path.as_deref()) {
+    match (request.method(), request.path_with_query().as_deref()) {
         (Method::Get, Some("/hash-all")) => {
             let urls = headers.iter().filter_map(|(k, v)| {
                 (k == "url")
                     .then_some(v)
-                    .and_then(|v| str::from_utf8(v).ok())
+                    .and_then(|v| std::str::from_utf8(v).ok())
                     .and_then(|v| Url::parse(v).ok())
             });
 
-            let results = urls.map(move |url| hash(url.clone()).map(move |result| (url, result)));
+            let results = urls.map(|url| async move {
+                let result = hash(&url).await;
+                (url, result)
+            });
 
             let mut results = stream::iter(results).buffer_unordered(MAX_CONCURRENCY);
 
@@ -38,19 +35,17 @@ async fn handle_request(request: IncomingRequest, response_out: ResponseOutparam
                 &Fields::new(&[("content-type".to_string(), b"text/plain".to_vec())]),
             );
 
-            let mut sink = executor::outgoing_response_body(&response);
+            let mut body = response.take_body();
 
             ResponseOutparam::set(response_out, Ok(response));
 
             while let Some((url, result)) = results.next().await {
-                sink.send(
-                    match result {
-                        Ok(hash) => format!("{url}: {hash}\n"),
-                        Err(e) => format!("{url}: {e:?}\n"),
-                    }
-                    .into_bytes(),
-                )
-                .await?;
+                let payload = match result {
+                    Ok(hash) => format!("{url}: {hash}\n"),
+                    Err(e) => format!("{url}: {e:?}\n"),
+                }
+                .into_bytes();
+                body.send(payload).await?;
             }
         }
 
@@ -59,21 +54,19 @@ async fn handle_request(request: IncomingRequest, response_out: ResponseOutparam
                 200,
                 &Fields::new(
                     &headers
-                        .iter()
-                        .filter_map(|(k, v)| {
-                            (k == "content-type").then_some((k.clone(), v.clone()))
-                        })
+                        .into_iter()
+                        .filter_map(|(k, v)| (k == "content-type").then_some((k, v)))
                         .collect::<Vec<_>>(),
                 ),
             );
 
-            let mut sink = executor::outgoing_response_body(&response);
+            let mut body = response.take_body();
 
             ResponseOutparam::set(response_out, Ok(response));
 
-            let mut stream = executor::incoming_request_body(request);
+            let mut stream = request.into_body();
             while let Some(chunk) = stream.try_next().await? {
-                sink.send(chunk).await?;
+                body.send(chunk).await?;
             }
         }
 
@@ -91,7 +84,7 @@ async fn handle_request(request: IncomingRequest, response_out: ResponseOutparam
     Ok(())
 }
 
-async fn hash(url: Url) -> Result<String> {
+async fn hash(url: &Url) -> Result<String> {
     let request = OutgoingRequest::new(
         &Method::Get,
         Some(url.path()),
@@ -104,7 +97,7 @@ async fn hash(url: Url) -> Result<String> {
         &Fields::new(&[]),
     );
 
-    let response = executor::outgoing_request_send(request).await?;
+    let response = send(request).await?;
 
     let status = response.status();
 
@@ -112,9 +105,10 @@ async fn hash(url: Url) -> Result<String> {
         bail!("unexpected status: {status}");
     }
 
-    let mut body = executor::incoming_response_body(response);
+    let mut body = response.into_body();
 
-    let mut hasher = Sha256::new();
+    use sha2::Digest;
+    let mut hasher = sha2::Sha256::new();
     while let Some(chunk) = body.try_next().await? {
         hasher.update(&chunk);
     }
